修正功能

+++ database/migrations/2016_01_02_050423_create_project_workitems_table.php
@@ -19,7 +19,7 @@ class CreateProjectWorkitemsTable extends Migration
             $table->integer('cost_type_id')->unsigned();
             $table->string('name');
             $table->integer('order')->unsigned();
-            $table->integer('amount')->unsigned();
+            $table->double('amount', 15, 8)->unsigned();
             $table->integer('unit_price')->unsigned();
             $table->timestamps();
         });
